requirements.txt
flask==2.2.5
scikit-learn==1.3.2
Werkzeug==3.0.0
python-multipart==0.0.5
nltk==3.8.1
tldextract
google.generativeai